app/requirements.txt
Flask
Flask-Cors
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
opencv-python
yt-dlp
loguru
python-dotenv
requests
werkzeug
